+++ src/pages/index.ts
@@ -1,5 +1,5 @@
 export { default as HomePage } from '@/pages/home';
-export { default as SignupPage } from '@/pages/sigunup';
+export { default as SignupPage } from '@/pages/signup';
 export { default as SearchResultPage } from '@/pages/searchResult';
 export { default as NotFoundPage } from '@/pages/notFound';
 export { default as MyPage } from '@/pages/myPage';
